+++ portfolio/src/test/java/com/google/sps/TimeRangeGroupTest.java
@@ -75,18 +75,6 @@ public class TimeRangeGroupTest {
     Assert.assertFalse(timeRangeGroup.hasTimeRange(timeRangeFive));
   }
 
-  /** Tests for the has method in an empty time range group. */
-  @Test
-  public void testHasRangeEmpty() {
-    List<TimeRange> timeRanges = Arrays.asList();
-    Instant timeRangeOneStart = Instant.now();
-    Instant timeRangeOneEnd = timeRangeOneStart.plusSeconds(1000);
-
-    TimeRange timeRangeOne = TimeRange.fromStartEnd(timeRangeOneStart, timeRangeOneEnd);
-
-    Assert.assertFalse(timeRangeGroup.hasTimeRange(timeRangeOne));
-  }
-
   /** Tests for merging two overlapping ranges. */
   @Test
   public void testMergeTwoRanges() {
@@ -312,6 +300,40 @@ public class TimeRangeGroupTest {
     Assert.assertEquals(expected, actual);
   }
 
+  /** Tests for adding one time range that overlaps with the existing one. */
+  @Test
+  public void testAddOneOverlappingRangeAndMore() {
+    // Time Ranges:       |-----A-----|      |----C----|
+    // To add:          |--------B--------|
+    // Result:          |-----------------|  |---------|
+
+    Instant timeRangeOneStart = Instant.now();
+    Instant timeRangeOneEnd = timeRangeOneStart.plusSeconds(1000);
+    Instant timeRangeTwoStart = timeRangeOneEnd.plusSeconds(3000);
+    Instant timeRangeTwoEnd = timeRangeTwoStart.plusSeconds(1000);
+    TimeRange timeRangeOne = TimeRange.fromStartEnd(timeRangeOneStart, timeRangeOneEnd);
+    TimeRange timeRangeTwo = TimeRange.fromStartEnd(timeRangeTwoStart, timeRangeTwoEnd);
+
+    List<TimeRange> originalTimeRanges = Arrays.asList(timeRangeOne, timeRangeTwo);
+    for (TimeRange timeRange : originalTimeRanges) {
+      timeRangeGroup.addTimeRange(timeRange);
+    }
+
+    Instant timeRangeToAddStart = timeRangeOneStart.minusSeconds(500);
+    Instant timeRangeToAddEnd = timeRangeOneEnd.plusSeconds(500);
+    TimeRange timeRangeToAdd = TimeRange.fromStartEnd(timeRangeToAddStart, timeRangeToAddEnd);
+    List<TimeRange> expected = Arrays.asList(timeRangeTwo, timeRangeToAdd);
+
+    timeRangeGroup.addTimeRange(timeRangeToAdd);
+    List<TimeRange> actual = new ArrayList();
+    for (TimeRange t : timeRangeGroup) {
+      actual.add(t);
+    }
+
+    Collections.sort(actual, TimeRange.SORT_BY_TIME_RANGE_DURATION_ASCENDING_THEN_START_TIME);
+    Assert.assertEquals(expected, actual);
+  }
+
   /** Tests for adding one time range that is contained by an existing one. */
   @Test
   public void testAddTimeRangeContainedByExisting() {
@@ -324,6 +346,9 @@ public class TimeRangeGroupTest {
     TimeRange timeRangeOne = TimeRange.fromStartEnd(timeRangeOneStart, timeRangeOneEnd);
 
     List<TimeRange> originalTimeRanges = Arrays.asList(timeRangeOne);
+    for (TimeRange timeRange : originalTimeRanges) {
+      timeRangeGroup.addTimeRange(timeRange);
+    }
 
     Instant timeRangeToAddStart = timeRangeOneStart.plusSeconds(100);
     Instant timeRangeToAddEnd = timeRangeOneEnd.minusSeconds(100);
@@ -331,12 +356,44 @@ public class TimeRangeGroupTest {
     List<TimeRange> expected = Arrays.asList(timeRangeOne);
 
     timeRangeGroup.addTimeRange(timeRangeToAdd);
-    Iterator<TimeRange> actualIterator = timeRangeGroup.iterator();
 
     List<TimeRange> actual = new ArrayList();
-    while (actualIterator.hasNext()) {
-      TimeRange newTimeRange = actualIterator.next();
-      actual.add(newTimeRange);
+    for (TimeRange timeRange : timeRangeGroup) {
+      actual.add(timeRange);
+    }
+
+    Collections.sort(actual, TimeRange.SORT_BY_TIME_RANGE_DURATION_ASCENDING_THEN_START_TIME);
+    Assert.assertEquals(expected, actual);
+  }
+
+  /** Tests for adding one time range that is contained by an existing one. */
+  @Test
+  public void testAddTimeRangeContainedByExistingPlus() {
+    // Time Ranges:  |-------A-------|      |----C----|
+    // To add:          |---B-----|
+    // Result:       |---------------|      |---------|
+
+    Instant timeRangeOneStart = Instant.now();
+    Instant timeRangeOneEnd = timeRangeOneStart.plusSeconds(1000);
+    Instant timeRangeTwoStart = timeRangeOneEnd.plusSeconds(3000);
+    Instant timeRangeTwoEnd = timeRangeTwoStart.plusSeconds(1000);
+    TimeRange timeRangeOne = TimeRange.fromStartEnd(timeRangeOneStart, timeRangeOneEnd);
+    TimeRange timeRangeTwo = TimeRange.fromStartEnd(timeRangeTwoStart, timeRangeTwoEnd);
+
+    List<TimeRange> originalTimeRanges = Arrays.asList(timeRangeOne, timeRangeTwo);
+    for (TimeRange timeRange : originalTimeRanges) {
+      timeRangeGroup.addTimeRange(timeRange);
+    }
+
+    Instant timeRangeToAddStart = timeRangeOneStart.plusSeconds(100);
+    Instant timeRangeToAddEnd = timeRangeOneEnd.minusSeconds(100);
+    TimeRange timeRangeToAdd = TimeRange.fromStartEnd(timeRangeToAddStart, timeRangeToAddEnd);
+    List<TimeRange> expected = Arrays.asList(timeRangeOne, timeRangeTwo);
+
+    timeRangeGroup.addTimeRange(timeRangeToAdd);
+    List<TimeRange> actual = new ArrayList();
+    for (TimeRange t : timeRangeGroup) {
+      actual.add(t);
     }
 
     Collections.sort(actual, TimeRange.SORT_BY_TIME_RANGE_DURATION_ASCENDING_THEN_START_TIME);
